move bind_interval logic to common
